Implemented Merging Functionality
Merge function takes two tables and a bool expr.
A bool expr can be made up of IndexedVars and == != && and || operators, all with () to reset precedence.
leftMerge will always merge, making it so the left table takes priority, and if its value is empty then the second table will fill in the value.